+++ src/SearchInput/index.js
@@ -8,20 +8,32 @@
  */
 
 import React, { Component, PropTypes } from 'react'
+import classnames from 'classnames'
+import controlledPropValidator from '../_shared/propValidator/controlled'
 import Button from '../Button'
 import ClearableInput from '../ClearableInput'
-import classnames from 'classnames'
-import './main.less'
+import './index.less'
 
 class SearchInput extends Component {
 
-  constructor() {
+  constructor(props) {
     super()
-    this.value = ''
+    this.state = {
+      value: 'value' in props ? props.value : props.defaultValue
+    }
   }
 
-  componentWillMount() {
-    this.value = this.props.defaultValue || ''
+  handleChange(value) {
+    this.setState({ value })
+    this.props.onChange && this.props.onChange(value)
+  }
+
+  handleSearch() {
+    this.props.onSearch && this.props.onSearch(this.state.value)
+  }
+
+  handleKeyUp(e) {
+    e.key === 'Enter' && this.handleSearch()
   }
 
   render() {
@@ -29,33 +41,23 @@ class SearchInput extends Component {
       className, label, size, defaultValue, onSearch, onChange, placeholder,
       ...other
     } = this.props
+    const { value } = this.state
+    delete other.value
     const width = this.props.width || '300px'
 
+    const inputProps = { value, size, placeholder }
+
     return (
-      <div className={classnames('bfd-search_input', className, size)} {...other}>
-        <ClearableInput style={{width}} defaultValue={this.value} size={size} onKeyUp={::this.handleKeyUp} onChange={::this.handleChange} inline placeholder={placeholder}/>
-        <Button size={size} onClick={::this.handleClick} icon="search">{label}</Button>
+      <div className={classnames('bfd-search_input', className)} {...other}>
+        <ClearableInput
+          style={{width}}
+          onKeyUp={::this.handleKeyUp}
+          onChange={::this.handleChange}
+          {...inputProps}
+        />
+        <Button size={size} onClick={::this.handleSearch} icon="search">{label}</Button>
       </div>
     )
-  }
-
-  handleChange(v) {
-    this.value = v
-    this.props.onChange && this.props.onChange(v)
-  }
-
-  handleClick() {
-    if (typeof this.props.onSearch == 'function') {
-      this.props.onSearch(this.value)
-    }
-  }
-
-  handleKeyUp(e) {
-    e.preventDefault()
-    const code = e.keyCode
-    if(code == 13) {
-      this.handleClick()
-    }
   }
 }
 
@@ -65,23 +67,26 @@ SearchInput.defaultProps = {
 
 SearchInput.propTypes = {
 
-  // 搜索框提示信息
-  placeholder: PropTypes.string,
+  // 输入框值
+  value: controlledPropValidator(PropTypes.oneOfType([PropTypes.string, PropTypes.number])),
 
-  // 搜索按钮名称，默认 `搜索`
-  label: PropTypes.string,
-
-  // 搜索按钮单击事件，value为搜索框输入值
-  onSearch: PropTypes.func.isRequired,
+  // 同 value，不可控
+  defaultValue: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
 
   // 搜索框值改变事件，value为搜索框输入值
   onChange: PropTypes.func,
 
+  // 搜索按钮单击事件，value为搜索框输入值
+  onSearch: PropTypes.func.isRequired,
+
+  // 搜索按钮名称，默认 `搜索`
+  label: PropTypes.string,
+
+  // 搜索框提示信息
+  placeholder: PropTypes.string,
+
   // 输入框高度尺寸，默认中等尺寸
   size: PropTypes.oneOf(['sm', 'lg']),
-
-  // 输入框默认值
-  defaultValue: PropTypes.string,
 
   // 输入框宽度，默认为300px
   width: PropTypes.string