+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Preview
+
+- SearchInput 新增可控属性 value
+- 因 [pace](http://github.hubspot.com/pace/docs/welcome/) 加载进度不够精确，文档站页面切换进度条升级到 [nprocess](http://ricostacruz.com/nprogress/)，与 react-router 配合使用
+
+
 ## v1.5.6
 
 *2017-01-06*